[Refactor] Separate Weekly Hot 5 component to prepare for server-side rendering
